Horizontal Toolbar is visible every tab in GridStudy
