+++ src/components/map-viewer.jsx
@@ -10,11 +10,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import { setStudyDisplayMode } from '../redux/actions';
 import { DRAW_EVENT, DRAW_MODES } from '@powsybl/network-viewer';
 import { ReactFlowProvider } from '@xyflow/react';
-import HorizontalToolbar from './horizontal-toolbar';
 import NetworkModificationTreePane from './network-modification-tree-pane';
 import NetworkMapTab from './network/network-map-tab';
 import { DiagramPane } from './diagrams/diagram-pane';
-import { darken } from '@mui/material/styles';
 
 import { Global, css } from '@emotion/react';
 import { EQUIPMENT_TYPES } from './utils/equipment-types';
@@ -36,11 +34,6 @@ const styles = {
         flexDirection: 'row',
         height: '100%',
     },
-    horizontalToolbar: (theme) => ({
-        backgroundColor: darken(theme.palette.background.paper, 0.2),
-        display: 'flex',
-        flexDirection: 'row',
-    }),
     table: {
         display: 'flex',
         flexDirection: 'column',
@@ -131,8 +124,6 @@ export const Content = () => (
 
 const MapViewer = ({
     studyUuid,
-    studyName,
-    studyPath,
     currentNode,
     currentRootNetworkUuid,
     view,
@@ -247,9 +238,6 @@ const MapViewer = ({
 
     return (
         <Box sx={styles.table}>
-            <Box sx={styles.horizontalToolbar}>
-                <HorizontalToolbar studyName={studyName} studyPath={studyPath} />
-            </Box>
             <Box sx={styles.mapAndTreeContainer}>
                 {/* Waiting for map geodata is unnecessary. The map has is proper loader implementation */}
                 {/* This WaitingLoader is placed here to block functionnalities, hiding under components with some opacity*/}
