+++ src/components/study-pane.jsx
@@ -45,7 +45,7 @@ const styles = {
     },
 };
 
-const StudyPane = ({ studyUuid, currentNode, studyName, studyPath, currentRootNetworkUuid, ...props }) => {
+const StudyPane = ({ studyUuid, currentNode, currentRootNetworkUuid, ...props }) => {
     const [tableEquipment, setTableEquipment] = useState({
         id: null,
         type: null,
@@ -80,8 +80,6 @@ const StudyPane = ({ studyUuid, currentNode, studyName, studyPath, currentRootNe
                     studyUuid={studyUuid}
                     currentNode={currentNode}
                     currentRootNetworkUuid={currentRootNetworkUuid}
-                    studyName={studyName}
-                    studyPath={studyPath}
                     view={props.view}
                     openDiagramView={openDiagramView}
                     tableEquipment={tableEquipment}
